Move `item-group` folder from `src/ui` to `src/`, update import paths
