+++ packages/components/src/item-group/item.ts
@@ -1,7 +1,7 @@
 /**
  * Internal dependencies
  */
-import { createComponent } from '../utils';
+import { createComponent } from '../ui/utils';
 import { useItem } from './use-item';
 
 export default createComponent( {
