+++ packages/components/src/item-group/use-item-group.ts
@@ -1,15 +1,15 @@
 /**
  * Internal dependencies
  */
-import { useContextSystem } from '../context';
+import { useContextSystem } from '../ui/context';
 // eslint-disable-next-line no-duplicate-imports
-import type { PolymorphicComponentProps } from '../context';
+import type { PolymorphicComponentProps } from '../ui/context';
 
 /**
  * Internal dependencies
  */
 import * as styles from './styles';
-import { useCx } from '../../utils/hooks/use-cx';
+import { useCx } from '../utils/hooks/use-cx';
 import type { ItemGroupProps } from './types';
 
 export function useItemGroup(
